requirements.txt
python-jobspy
anthropic
openai
pdfminer.six
python-dotenv
pyyaml
jinja2
requests
